refactor: StorageContextをstorageディレクトリに移動 関連インポートを更新
