+++ app/page.tsx
@@ -2,7 +2,7 @@
 
 import { PanelProvider } from "@/app/feature/splitter/context/PanelContext";
 import { SelectedDataProvider } from "@/app/context/AppContext";
-import { LocalStorageProvider } from "@/app/feature/localStorage/context/localStorageContext";
+import { LocalStorageProvider } from "@/app/feature/storage/context/StorageContext";
 import Header from "@/app/components/ui/header/Header";
 import Main from "@/app/components/ui/main/Main";
